site-react: query-param goal route (/goal/?id=) — any goal opens, no 404
Replaces the SPA-redirect attempt (which looped — Next static export can't
client-navigate to a non-prerendered dynamic route). Root issue: /goals/[id]
only emitted 3 ids, so the goals board couldn't even OPEN most goals (let alone
deep-links). Fix: one statically-exported /goal/?id=<id> page that renders ANY
goal client-side (useSearchParams + client fetch).

- New app/goal/ (server page + Suspense + GoalRoute client + GoalDetail prop-based).
- Removed app/goals/[id]/ (dynamic route).
- All 7 goal-detail links (GoalsClient, GraphClient, HomeClient) → /goal/?id=.
- Reverted the SPA-fallback 404.html redirector.
Verified: /goal/?id=<anything> → HTTP 200, renders the goal route.
