+++ WebSite/site-react/app/goal/GoalRoute.tsx
@@ -1,0 +1,9 @@
+"use client";
+
+import { useSearchParams } from "next/navigation";
+import GoalDetail from "./_components/GoalDetail";
+
+export default function GoalRoute() {
+  const id = useSearchParams().get("id") ?? "";
+  return <GoalDetail id={id} />;
+}
